+++ Changes.md
@@ -23,6 +23,7 @@ Fixes
 - PlugAlgo : Fixed metadata handling when promoting plugs which were themselves promoted from the constructor of a custom node.
 - NodeAlgo : Fixed corner case where a preset could be listed twice if it was specified by both `preset:<name>` and `presetNames`.
 - GafferUI : Fixed incorrect Tree/TableView horizontal header height.
+- Spreadsheet : Fixed bug that allowed values to be pasted onto plugs with inputs.
 
 API
 ---